+++ hive/lib/src/backend/js/native/storage_backend_js.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:html';
 import 'dart:indexed_db';
+import 'dart:js' as js;
 import 'dart:js_util';
 import 'dart:typed_data';
 
@@ -18,11 +19,12 @@ class StorageBackendJs extends StorageBackend {
   static const _bytePrefix = [0x90, 0xA9];
   final Database _db;
   final HiveCipher? _cipher;
+  final String objectStoreName;
 
   TypeRegistry _registry;
 
   /// Not part of public API
-  StorageBackendJs(this._db, this._cipher,
+  StorageBackendJs(this._db, this._cipher, this.objectStoreName,
       [this._registry = TypeRegistryImpl.nullImpl]);
 
   @override
@@ -95,10 +97,10 @@ class StorageBackendJs extends StorageBackend {
 
   /// Not part of public API
   @visibleForTesting
-  ObjectStore getStore(bool write, [String box = 'box']) {
+  ObjectStore getStore(bool write) {
     return _db
-        .transaction(box, write ? 'readwrite' : 'readonly')
-        .objectStore(box);
+        .transaction(objectStoreName, write ? 'readwrite' : 'readonly')
+        .objectStore(objectStoreName);
   }
 
   /// Not part of public API
@@ -199,6 +201,12 @@ class StorageBackendJs extends StorageBackend {
 
   @override
   Future<void> deleteFromDisk() {
-    return window.indexedDB!.deleteDatabase(_db.name!);
+    final indexDB = js.context.hasProperty('window')
+        ? window.indexedDB
+        : WorkerGlobalScope.instance.indexedDB;
+    return indexDB!.deleteDatabase(_db.name!);
   }
+
+  @override
+  Future<void> flush() => Future.value();
 }