+++ hive/lib/src/box/lazy_box_impl.dart
@@ -5,8 +5,8 @@ import 'package:hive/hive.dart';
 import 'package:hive/src/backend/storage_backend.dart';
 import 'package:hive/src/binary/frame.dart';
 import 'package:hive/src/box/box_base_impl.dart';
-import 'package:hive/src/object/hive_object.dart';
 import 'package:hive/src/hive_impl.dart';
+import 'package:hive/src/object/hive_object.dart';
 
 /// Not part of public API
 class LazyBoxImpl<E> extends BoxBaseImpl<E> implements LazyBox<E> {
@@ -30,12 +30,12 @@ class LazyBoxImpl<E> extends BoxBaseImpl<E> implements LazyBox<E> {
 
     if (frame != null) {
       var value = await backend.readValue(frame);
-      if (value is HiveObject) {
+      if (value is HiveObjectMixin) {
         value.init(key, this);
       }
       return value as E?;
     } else {
-      if (defaultValue != null && defaultValue is HiveObject) {
+      if (defaultValue != null && defaultValue is HiveObjectMixin) {
         defaultValue.init(key, this);
       }
       return defaultValue;
@@ -63,10 +63,10 @@ class LazyBoxImpl<E> extends BoxBaseImpl<E> implements LazyBox<E> {
     await backend.writeFrames(frames);
 
     for (var frame in frames) {
-      if (frame.value is HiveObject) {
-        (frame.value as HiveObject).init(frame.key, this);
+      if (frame.value is HiveObjectMixin) {
+        (frame.value as HiveObjectMixin).init(frame.key, this);
       }
-      keystore.insert(frame.toLazy());
+      keystore.insert(frame, lazy: true);
     }
 
     await performCompactionIfNeeded();
@@ -91,5 +91,10 @@ class LazyBoxImpl<E> extends BoxBaseImpl<E> implements LazyBox<E> {
     }
 
     await performCompactionIfNeeded();
+  }
+
+  @override
+  Future<void> flush() async {
+    await backend.flush();
   }
 }