+++ hive/lib/src/hive.dart
@@ -5,7 +5,11 @@ abstract class HiveInterface implements TypeRegistry {
   /// Initialize Hive by giving it a home directory.
   ///
   /// (Not necessary in the browser)
-  void init(String path);
+  void init(
+    String? path, {
+    HiveStorageBackendPreference backendPreference =
+        HiveStorageBackendPreference.native,
+  });
 
   /// Opens a box.
   ///
@@ -19,7 +23,8 @@ abstract class HiveInterface implements TypeRegistry {
     bool crashRecovery = true,
     String? path,
     Uint8List? bytes,
-    @deprecated List<int>? encryptionKey,
+    String? collection,
+    @Deprecated('Use encryptionCipher instead') List<int>? encryptionKey,
   });
 
   /// Opens a lazy box.
@@ -33,7 +38,8 @@ abstract class HiveInterface implements TypeRegistry {
     CompactionStrategy compactionStrategy = defaultCompactionStrategy,
     bool crashRecovery = true,
     String? path,
-    @deprecated List<int>? encryptionKey,
+    String? collection,
+    @Deprecated('Use encryptionCipher instead') List<int>? encryptionKey,
   });
 
   /// Returns a previously opened box.
@@ -51,7 +57,7 @@ abstract class HiveInterface implements TypeRegistry {
   /// Removes the file which contains the box and closes the box.
   ///
   /// In the browser, the IndexedDB database is being removed.
-  Future<void> deleteBoxFromDisk(String name);
+  Future<void> deleteBoxFromDisk(String name, {String? path});
 
   /// Deletes all currently open boxes from disk.
   ///
@@ -62,7 +68,7 @@ abstract class HiveInterface implements TypeRegistry {
   List<int> generateSecureKey();
 
   /// Checks if a box exists
-  Future<bool> boxExists(String name);
+  Future<bool> boxExists(String name, {String? path});
 }
 
 ///