+++ hive/lib/src/object/hive_object.dart
@@ -1,14 +1,14 @@
 library hive_object_internal;
 
 import 'package:hive/hive.dart';
-import 'package:meta/meta.dart';
 import 'package:hive/src/object/hive_list_impl.dart';
+import 'package:meta/meta.dart';
 
 part 'hive_object_internal.dart';
 
 /// Extend `HiveObject` to add useful methods to the objects you want to store
 /// in Hive
-abstract class HiveObject {
+mixin HiveObjectMixin {
   BoxBase? _box;
 
   dynamic _key;
@@ -57,3 +57,5 @@ abstract class HiveObject {
     return false;
   }
 }
+
+abstract class HiveObject with HiveObjectMixin {}