+++ hive/test/integration/hive_object_test.dart
@@ -1,16 +1,18 @@
 import 'package:hive/hive.dart';
-import 'package:hive/src/hive_impl.dart';
 import 'package:test/test.dart';
 
 import 'integration.dart';
 
-class _TestObject extends HiveObject {
+class _TestObject with HiveObjectMixin {
   String name;
 
   _TestObject(this.name);
 
   @override
   bool operator ==(dynamic other) => other is _TestObject && other.name == name;
+
+  @override
+  int get hashCode => runtimeType.hashCode ^ name.hashCode;
 }
 
 class _TestObjectAdapter extends TypeAdapter<_TestObject> {
@@ -29,8 +31,8 @@ class _TestObjectAdapter extends TypeAdapter<_TestObject> {
 }
 
 Future _performTest(bool lazy) async {
-  var hive = HiveImpl();
-  hive.registerAdapter(_TestObjectAdapter());
+  var hive = await createHive();
+  hive.registerAdapter<_TestObject>(_TestObjectAdapter());
   var box = await openBox(lazy, hive: hive);
 
   var obj1 = _TestObject('test1');