+++ hive/test/tests/box/keystore_test.dart
@@ -1,10 +1,6 @@
-import 'package:hive/hive.dart';
 import 'package:hive/src/binary/frame.dart';
-import 'package:hive/src/box/change_notifier.dart';
 import 'package:hive/src/box/keystore.dart';
-import 'package:hive/src/object/hive_object.dart';
-import 'package:mockito/annotations.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 
 import '../mocks.dart';
@@ -256,10 +252,10 @@ void main() {
           var keystore = Keystore.debug(notifier: notifier);
 
           keystore.insert(Frame('key1', 'val1'));
-          verify(notifier.notify(Frame('key1', 'val1')));
+          verify(() => notifier.notify(Frame('key1', 'val1')));
 
           keystore.insert(Frame('key1', 'val2'));
-          verify(notifier.notify(Frame('key1', 'val2')));
+          verify(() => notifier.notify(Frame('key1', 'val2')));
         });
       });
 
@@ -311,7 +307,7 @@ void main() {
           reset(notifier);
 
           keystore.insert(Frame.deleted('key1'));
-          verify(notifier.notify(Frame.deleted('key1')));
+          verify(() => notifier.notify(Frame.deleted('key1')));
 
           keystore.insert(Frame.deleted('key1'));
           verifyNoMoreInteractions(notifier);
@@ -332,8 +328,8 @@ void main() {
         expect(created, true);
         expect(keystore.transactions.first.added, ['key1', 'key2']);
         expect(keystore.frames, [Frame('key1', 'val1'), Frame('key2', 'val2')]);
-        verify(notifier.notify(Frame('key1', 'val1')));
-        verify(notifier.notify(Frame('key2', 'val2')));
+        verify(() => notifier.notify(Frame('key1', 'val1')));
+        verify(() => notifier.notify(Frame('key2', 'val2')));
       });
 
       test('overriding existing keys', () {
@@ -354,8 +350,8 @@ void main() {
           'key1': Frame('key1', 'val1'),
         });
         expect(keystore.frames, [Frame('key1', 'val2'), Frame('key2', 'val3')]);
-        verify(notifier.notify(Frame('key1', 'val2')));
-        verify(notifier.notify(Frame('key2', 'val3')));
+        verify(() => notifier.notify(Frame('key1', 'val2')));
+        verify(() => notifier.notify(Frame('key2', 'val3')));
       });
 
       test('empty transaction', () {
@@ -394,7 +390,7 @@ void main() {
           'key1': Frame('key1', 'val1'),
         });
         expect(keystore.frames, [Frame('key2', 'val2')]);
-        verify(notifier.notify(Frame.deleted('key1')));
+        verify(() => notifier.notify(Frame.deleted('key1')));
       });
     });
 
@@ -433,7 +429,7 @@ void main() {
           keystore.transactions,
           [KeyTransaction()..added.add('otherKey')],
         );
-        verify(notifier.notify(Frame.deleted('key')));
+        verify(() => notifier.notify(Frame.deleted('key')));
       });
 
       test('add then override', () {
@@ -486,7 +482,7 @@ void main() {
         keystore.cancelTransaction();
         expect(keystore.frames, [Frame('key', 'val1')]);
         expectTrx(keystore.transactions, []);
-        verify(notifier.notify(Frame('key', 'val1')));
+        verify(() => notifier.notify(Frame('key', 'val1')));
       });
 
       test('override then add', () {
@@ -539,7 +535,7 @@ void main() {
         keystore.cancelTransaction();
         expect(keystore.frames, [Frame('key', 'val1')]);
         expectTrx(keystore.transactions, []);
-        verify(notifier.notify(Frame('key', 'val1')));
+        verify(() => notifier.notify(Frame('key', 'val1')));
       });
 
       test('delete then add', () {
@@ -616,8 +612,8 @@ void main() {
         reset(notifier);
 
         keystore.clear();
-        verify(notifier.notify(Frame.deleted('key1')));
-        verify(notifier.notify(Frame.deleted('key2')));
+        verify(() => notifier.notify(Frame.deleted('key1')));
+        verify(() => notifier.notify(Frame.deleted('key2')));
       });
     });
   });