+++ hive_flutter/test/adapters/color_adapter_test.dart
@@ -1,8 +1,8 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:hive/hive.dart';
-import 'package:hive_flutter/src/adapters/color_adapter.dart';
+import 'dart:ui' show Color;
+
+import 'package:hive_flutter/adapters.dart';
 import 'package:mockito/mockito.dart';
+import 'package:test/test.dart';
 
 import '../mocks.dart';