+++ hive_generator/lib/hive_generator.dart
@@ -2,4 +2,5 @@ import 'package:build/build.dart';
 import 'package:hive_generator/src/type_adapter_generator.dart';
 import 'package:source_gen/source_gen.dart';
 
-Builder getBuilder(BuilderOptions options) => PartBuilder([TypeAdapterGenerator()], '.hive.dart');
+Builder getBuilder(BuilderOptions options) =>
+    SharedPartBuilder([TypeAdapterGenerator()], 'hive_generator');