+++ hive_generator/CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.1.3
+
+- Updated `analyzer` version constraints to `<5.0.0`
+
+## 1.1.2
+
+- Updated `analyzer` version constraints
+
+## 1.1.1
+
+- Updated dependency versions
+
+## 1.1.0
+
+- Fixes empty class generates adapter with warnings - [#638](https://github.com/hivedb/hive/issues/638)
+- Default value support for class type adapter generators - [#625](https://github.com/hivedb/hive/issues/625)
+
+## 1.0.1
+
+- Fixed dependencies for null-safety compatibility
+
 ## 1.0.0
 
 - Stable null-safety version